fix(extras): kitty base16 correction on brights (color8) (#110)
* fix(kitty.conf): base16 correction on brights (color8)

If color0 and color8 are the same, the zsh-autosuggestions are the same colour as the background making it difficult to read.

I used the wezterm config as reference and found that the color8 used in this config was incorrect thus fixed.

* fix(kitty.lua): base16 correction in the right location this time

Edited kitty.lua so that it won't overwrite the .conf file.